gpt 연결 코드 파일 분리
